fix(test): decrypt captured frames without reparsing their length
`CapturingMockTransport::sent()` already splits each write into its frames,
prefix intact, so the helper was walking a length prefix that had been walked
for it and adding a third copy of that parse to the crate. It now checks the
declared length against the entry instead, which turns a caller that passes
raw writes into a failed assertion rather than a decrypt of garbage.

The noise socket the client-test helper installs now reports to the client's
stats as well as its tap, matching the real socket and the other fixture: a
test that later asserts on `frames_sent` would have read zero and blamed the
code under test.

The allocation delta is compared as a signed difference. Both sides are
empirical minima off a process-global counter, and an inversion has to print
the two numbers rather than panic inside the subtraction.

Also fixes the observability doc: it claimed five distinct send paths and then
listed callers that funnel through two, and still described VoIP sockets as
passing `None`, which this branch replaced with `SendObservers::default()`.

+++ src/socket/noise_socket.rs
@@ -2061,8 +2061,10 @@ mod tests {
             idle <= 4,
             "an unwatched send must cost what it always did, got {idle}"
         );
+        // Signed: both are empirical minima off a process-global counter, and an
+        // inversion has to report the numbers rather than panic on underflow.
         assert_eq!(
-            watched - idle,
+            watched as i64 - idle as i64,
             2,
             "watching must cost a constant per frame (idle {idle}, watched {watched})"
         );
